// front-end-app/src/app/model/category.model.ts
export enum CategoryModel {
  PersonalCat = 'personal_cat',
  MotorcycleCat = 'motorcycle_cat',
  ProviderCat = 'provider_cat',
  TruckCat = 'truck_cat',
  BuildCat = 'build_cat',
  AgriculturalCat = 'agricultural_cat',
  TrailerCat = 'trailer_cat',
  BusCat = 'bus_cat',
  ShipCat = 'ship_cat',
  PlaneCat = 'plane_cat',
  PartCat = 'part_cat',
}